move repl stuff to a REPL object in rainier-plot
