+++ rainier-core/src/main/scala/com/stripe/rainier/sampler/RNG.scala
@@ -11,7 +11,7 @@ trait RNG {
 }
 
 object RNG {
-  lazy val default: RNG = {
+  implicit lazy val default: RNG = {
     val seed = System.currentTimeMillis
     ScalaRNG(seed)
   }
