+++ rainier-plot/src/main/scala/com/stripe/rainier/plot/REPL.scala
@@ -1,21 +1,8 @@
-package com.stripe.rainier
+package com.stripe.rainier.plot
 
-import com.stripe.rainier.sampler._
 import java.io._
 
-package object repl {
-  def plot1D[N](seq: Seq[N])(implicit num: Numeric[N]): Unit = {
-    println(DensityPlot().plot1D(seq.map(num.toDouble)).mkString("\n"))
-  }
-
-  def plot2D[M, N](seq: Seq[(M, N)])(implicit n: Numeric[N],
-                                     m: Numeric[M]): Unit = {
-    println(
-      DensityPlot()
-        .plot2D(seq.map { case (a, b) => (m.toDouble(a), n.toDouble(b)) })
-        .mkString("\n"))
-  }
-
+object REPL {
   def writeCSV(path: String, seq: Seq[Map[String, Double]]): Unit = {
     val fieldNames = seq.map(_.keys.toSet).reduce(_ ++ _).toList
     val pw = new PrintWriter(new File(path))
@@ -189,5 +176,4 @@ package object repl {
   private def leftPad(s: String, len: Int, elem: Char): String =
     s.reverse.padTo(len, elem).reverse
 
-  implicit val rng: RNG = RNG.default
 }
